change from vec to btreeset
