+++ crates/katana/pool/src/ordering.rs
@@ -76,15 +76,15 @@ impl PartialEq for TxSubmissionNonce {
 /// This ordering implementation uses the transaction's tip as the priority value. We don't have a
 /// use case for this ordering implementation yet, but it's mostly used for testing.
 #[derive(Debug)]
-pub struct Tip<T>(PhantomData<T>);
+pub struct TipOrdering<T>(PhantomData<T>);
 
-impl<T> Tip<T> {
+impl<T> TipOrdering<T> {
     pub fn new() -> Self {
         Self(PhantomData)
     }
 }
 
-impl<T: PoolTransaction> PoolOrd for Tip<T> {
+impl<T: PoolTransaction> PoolOrd for TipOrdering<T> {
     type Transaction = T;
     type PriorityValue = u64;
 
@@ -93,8 +93,74 @@ impl<T: PoolTransaction> PoolOrd for Tip<T> {
     }
 }
 
-impl<T> Default for Tip<T> {
+impl<T> Default for TipOrdering<T> {
     fn default() -> Self {
         Self::new()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+    use crate::ordering::{self, FiFo};
+    use crate::pool::test_utils::*;
+    use crate::tx::PoolTransaction;
+    use crate::validation::NoopValidator;
+    use crate::{Pool, TransactionPool};
+
+    #[test]
+    fn fifo_ordering() {
+        // Create mock transactions
+        let txs = [PoolTx::new(), PoolTx::new(), PoolTx::new(), PoolTx::new(), PoolTx::new()];
+
+        // Create a pool with FiFo ordering
+        let pool = Pool::new(NoopValidator::new(), FiFo::new());
+
+        // Add transactions to the pool
+        txs.iter().for_each(|tx| {
+            let _ = pool.add_transaction(tx.clone());
+        });
+
+        // Get pending transactions
+        let pendings = pool.take_transactions().collect::<Vec<_>>();
+
+        // Assert that the transactions are in the order they were added (first to last)
+        pendings.iter().zip(txs).for_each(|(pending, tx)| {
+            assert_eq!(pending.tx.as_ref(), &tx);
+        });
+    }
+
+    #[test]
+    fn tip_based_ordering() {
+        // Create mock transactions with different tips and in random order
+        let txs = [
+            PoolTx::new().with_tip(1),
+            PoolTx::new().with_tip(6),
+            PoolTx::new().with_tip(3),
+            PoolTx::new().with_tip(2),
+            PoolTx::new().with_tip(5),
+            PoolTx::new().with_tip(4),
+            PoolTx::new().with_tip(7),
+        ];
+
+        // Create a pool with tip-based ordering
+        let pool = Pool::new(NoopValidator::new(), ordering::TipOrdering::new());
+
+        // Add transactions to the pool
+        txs.iter().for_each(|tx| {
+            let _ = pool.add_transaction(tx.clone());
+        });
+
+        // Get pending transactions
+        let pending = pool.take_transactions().collect::<Vec<_>>();
+
+        // Assert that the transactions are ordered by tip (highest to lowest)
+        assert_eq!(pending[0].tx.tip(), 7);
+        assert_eq!(pending[1].tx.tip(), 6);
+        assert_eq!(pending[2].tx.tip(), 5);
+        assert_eq!(pending[3].tx.tip(), 4);
+        assert_eq!(pending[4].tx.tip(), 3);
+        assert_eq!(pending[5].tx.tip(), 2);
+        assert_eq!(pending[6].tx.tip(), 1);
     }
 }
